llava_model/requirement.txt

torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
pillow>=9.0.0
sentencepiece>=0.1.99
protobuf>=3.20.0
llava-torch>=1.1.1
scipy>=1.11.0
einops>=0.7.0
gradio 